More 64-bit type updates

+++ Source/Clients/Spreadsheets/GDataEntryWorksheet.m
@@ -1,17 +1,17 @@
 /* Copyright (c) 2007 Google Inc.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 //
 //  GDataEntryWorksheet.m
@@ -22,9 +22,6 @@
 #import "GDataEntryWorksheet.h"
 #import "GDataSpreadsheetConstants.h"
 #import "GDataRowColumnCount.h"
-
-// extensions
-
 
 
 @implementation GDataEntryWorksheet
@@ -62,7 +59,6 @@
   [super addExtensionDeclarations];
   
   Class entryClass = [self class];
-  
   
   // Worksheet extensions
   [self addExtensionDeclarationForParentClass:entryClass
@@ -105,24 +101,24 @@
 
 #pragma mark -
 
-- (int)rowCount {
+- (NSInteger)rowCount {
   GDataRowCount *rowCount = [self objectForExtensionClass:[GDataRowCount class]];
   
   return [rowCount count];
 }
 
-- (void)setRowCount:(int)val {
+- (void)setRowCount:(NSInteger)val {
   GDataRowCount *obj = [GDataRowCount rowCountWithInt:val];
   [self setObject:obj forExtensionClass:[GDataRowCount class]];
 }
 
-- (int)columnCount {
+- (NSInteger)columnCount {
   GDataColumnCount *columnCount = [self objectForExtensionClass:[GDataColumnCount class]];
   
   return [columnCount count];
 }
 
-- (void)setColumnCount:(int)val {
+- (void)setColumnCount:(NSInteger)val {
   GDataColumnCount *obj = [GDataColumnCount columnCountWithInt:val];
   [self setObject:obj forExtensionClass:[GDataColumnCount class]];
 }
